Updated schemas
