+++ app/models/incidents.py
@@ -34,6 +34,7 @@ class Incident(Base):
     was_transported_ambulance: Mapped[str] = mapped_column(String, nullable=True)
     ambulance_to_where: Mapped[str] = mapped_column(String, nullable=True)
     type_of_incident: Mapped[TypeOfIncident] = mapped_column(Enum(TypeOfIncident, name="type_of_incident_enum", native_enum=False), nullable=True)
+    location_of_incident: Mapped[str] = mapped_column(String, nullable=True)
     
     # SAMPLE answers
     signs_symptoms: Mapped[str] = mapped_column(String, nullable=True)
