+++ app/schemas/incident_schemas.py
@@ -1,6 +1,4 @@
-from uuid import UUID
-from pydantic import BaseModel, Field, model_validator, ConfigDict
-from datetime import datetime
+from pydantic import BaseModel, model_validator, ConfigDict
 from typing import Optional, Literal
 from enum import Enum
 
@@ -23,34 +21,6 @@ class TypeofInjury(Enum):
     CutScrape = "CutScrape"
     HeartAttack = "HeartAttack"
     Other = "Other"
-
-class LifeguardPayload(BaseModel):
-    name: str = Field(min_length=1)
-    phone: str = Field(min_length=12, max_length=14) # 10 DLC with up to 4 for +country code
-    region: str
-
-class Lifeguard(LifeguardPayload):
-    id: UUID
-    created: datetime
-
-    class Config:
-        from_attributes = True
-
-
-class Region(BaseModel):
-    id: UUID
-    name: str = Field(min_length=1)
-    created: datetime
-
-    class Config:
-        from_attributes = True
-
-class Manager(BaseModel):
-    id: UUID
-    name: str = Field(min_length=1)
-
-    class Config:
-        from_attributes = True
 
 class IncidentSummary(BaseModel):
     body_part_afflicted: Optional[str] = None
